refactor: 테스트용 fake input, output view 분리
